backend/app/auth.py
import os
from datetime import datetime, timedelta
from jose import jwt

SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret")
ALGORITHM = "HS256"


def create_access_token(subject: str, expires_delta: int = 60 * 24):
    to_encode = {"sub": subject, "exp": datetime.utcnow() + timedelta(minutes=expires_delta)}
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def verify_token(token: str):
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except Exception:
        return None
